edited the PR notes
