+++ src/components/Home/Home.css
@@ -1,0 +1,41 @@
+.profile {
+    display: flex;
+    align-items: flex-start;
+    justify-content: space-between;
+    margin: 50px 20px 0px;
+    padding-bottom: 50px;
+    text-align: start;
+    font-size: 20px;
+    padding-bottom: 50px;
+}
+
+.profile-title {
+    width: 50%;
+}
+
+.profile-title h1 {
+    font-size: 40px;
+    font-weight: 700;
+    color: gray;
+}
+
+.profile-title p {
+    font-weight: 700;
+    font-size: 25px;
+}
+
+.profile-description {
+    width: 50%;
+}
+
+.profile-description p {
+    width: 95%;
+    line-height: 1.9;
+    font-weight: 500;
+}
+
+.profile-description strong {
+    color: gray;
+    font-weight: 800;
+    font-size: 1.3em;
+}
